Upgrade 13.0.20.36, full clean images sync, no more duplication on post and import

+++ models/product.py
@@ -626,19 +626,22 @@ class product_product(models.Model):
         product = self
         ml_pics = {}
         for ix in range(0,len(pictures)):
-            ml_pics[pictures[ix]['id']] = True
+            ml_imgid = pictures[ix]['id']
+            if (ml_imgid):
+                ml_pics[ml_imgid] = True
+                _logger.info(ml_imgid)
 
-            duplicates = self.env["product.image"].search([('meli_imagen_id','=',pictures[ix]['id']),('product_tmpl_id','=',product_template.id)])
-            if (duplicates and len(duplicates)>1):
-                _logger.info("Removing template duplicates for "+str(pictures[ix]['id'])+" :"+str(len(duplicates)-1))
-                for ix in range(1,len(duplicates)):
-                    duplicates[ix].unlink()
+                duplicates = self.env["product.image"].search([('meli_imagen_id','=',ml_imgid),('product_tmpl_id','=',product_template.id)])
+                if (duplicates and len(duplicates)>1):
+                    _logger.info("Removing template duplicates for "+str(ml_imgid)+" :"+str(len(duplicates)-1))
+                    for ix in range(1,len(duplicates)):
+                        duplicates[ix].unlink()
 
-            duplicates = self.env["product.image"].search([('meli_imagen_id','=',pictures[ix]['id']),('product_variant_id','=',product.id)])
-            if (duplicates and len(duplicates)>1):
-                _logger.info("Removing variant duplicates for "+str(pictures[ix]['id'])+" :"+str(len(duplicates)-1))
-                for ix in range(1,len(duplicates)):
-                    duplicates[ix].unlink()
+                duplicates = self.env["product.image"].search([('meli_imagen_id','=',ml_imgid),('product_variant_id','=',product.id)])
+                if (duplicates and len(duplicates)>1):
+                    _logger.info("Removing variant duplicates for "+str(ml_imgid)+" :"+str(len(duplicates)-1))
+                    for ix in range(1,len(duplicates)):
+                        duplicates[ix].unlink()
 
         _logger.info(ml_pics)
